add refinement network
