+++ hilbert_predictor/data.py
@@ -3,12 +3,10 @@ import pickle
 import numpy as np
 import os
 import json
-import numpy as np
-import torch
 import random
 
+from .gilbert2d import flatten_2d_to_1d, unflatten_1d_to_2d
 from .args import args
-from .encoder import PositionEncoder, NUM_ENCODING_DIMENSIONS
 
 # Token definitions
 PAD_TOKEN = 10
@@ -24,11 +22,11 @@ MASK = 19
 NUM_TOKENS = END_SEQUENCE_TOKEN + 1
 
 if args.simple:
-    MAX_CONTEXT_LENGTH = 256
-    MAX_PREDICTION_LENGTH = 32
+    MAX_CONTEXT_LENGTH = 512
+    MAX_PREDICTION_LENGTH = 64
 else:
     MAX_CONTEXT_LENGTH = 1024
-    MAX_PREDICTION_LENGTH = 128
+    MAX_PREDICTION_LENGTH = 256
 
 MAX_SEQUENCE_LENGTH = MAX_CONTEXT_LENGTH + MAX_PREDICTION_LENGTH
 
@@ -37,126 +35,6 @@ evaluating_data = None
 simple_dataset = "arc-datasets/datasets/bitdata/"
 
 
-# Gilbert2D - Generalized Hilbert Curve for 2D space-filling
-def gilbert2d(width, height):
-    """
-    Generalized Hilbert ('gilbert') space-filling curve for arbitrary-sized
-    2D rectangular grids. Generates discrete 2D coordinates to fill a rectangle
-    of size (width x height).
-    """
-
-    if width >= height:
-        yield from generate2d(0, 0, width, 0, 0, height)
-    else:
-        yield from generate2d(0, 0, 0, height, width, 0)
-
-
-def sgn(x):
-    return -1 if x < 0 else (1 if x > 0 else 0)
-
-
-def generate2d(x, y, ax, ay, bx, by):
-
-    w = abs(ax + ay)
-    h = abs(bx + by)
-
-    (dax, day) = (sgn(ax), sgn(ay))  # unit major direction
-    (dbx, dby) = (sgn(bx), sgn(by))  # unit orthogonal direction
-
-    if h == 1:
-        # trivial row fill
-        for i in range(0, w):
-            yield (x, y)
-            (x, y) = (x + dax, y + day)
-        return
-
-    if w == 1:
-        # trivial column fill
-        for i in range(0, h):
-            yield (x, y)
-            (x, y) = (x + dbx, y + dby)
-        return
-
-    (ax2, ay2) = (ax // 2, ay // 2)
-    (bx2, by2) = (bx // 2, by // 2)
-
-    w2 = abs(ax2 + ay2)
-    h2 = abs(bx2 + by2)
-
-    if 2 * w > 3 * h:
-        if (w2 % 2) and (w > 2):
-            # prefer even steps
-            (ax2, ay2) = (ax2 + dax, ay2 + day)
-
-        # long case: split in two parts only
-        yield from generate2d(x, y, ax2, ay2, bx, by)
-        yield from generate2d(x + ax2, y + ay2, ax - ax2, ay - ay2, bx, by)
-
-    else:
-        if (h2 % 2) and (h > 2):
-            # prefer even steps
-            (bx2, by2) = (bx2 + dbx, by2 + dby)
-
-        # standard case: one step up, one long horizontal, one step down
-        yield from generate2d(x, y, bx2, by2, ax2, ay2)
-        yield from generate2d(x + bx2, y + by2, ax, ay, bx - bx2, by - by2)
-        yield from generate2d(
-            x + (ax - dax) + (bx2 - dbx),
-            y + (ay - day) + (by2 - dby),
-            -bx2,
-            -by2,
-            -(ax - ax2),
-            -(ay - ay2),
-        )
-
-
-def flatten_2d_to_1d(array):
-    array = np.array(array)
-    if array.ndim == 1:
-        return array.tolist()
-    elif array.ndim == 2:
-        height, width = array.shape
-        if height == 1:
-            return array[0].tolist()
-        elif width == 1:
-            return array.T[0].tolist()
-        else:
-            # 2D matrix: flatten using Hilbert curve
-            array_1d = [None] * (width * height)
-            for idx, (x, y) in enumerate(gilbert2d(width, height)):
-                array_1d[idx] = array[y][x]
-            return array_1d
-    else:
-        raise ValueError(f"Input array must be 1D or 2D, got {array.ndim}D")
-
-def create_mapping_table():
-    mapping_table = {}
-    position_encoder = PositionEncoder(30, 30)
-    for height in range(1, 31):
-        for width in range(1, 31):
-            key = (height, width)
-            if height == 1 or width == 1:
-                mapping = [(0, i) if height == 1 else (i, 0) for i in range(max(height, width))]
-            else:
-                mapping = list(gilbert2d(width, height))
-            encodings = position_encoder.compute_encodings(height, width)
-            mapping_table[key] = (mapping, encodings)
-    return mapping_table
-
-def save_mapping_table(mapping_table, filename='mapping_table.pkl'):
-    with open(filename, 'wb') as f:
-        pickle.dump(mapping_table, f)
-
-def load_mapping_table(filename='mapping_table.pkl'):
-    with open(filename, 'rb') as f:
-        return pickle.load(f)
-
-def unflatten_1d_to_2d(array_1d, width, height):
-    array_2d = [[None] * width for _ in range(height)]
-    for idx, (x, y) in enumerate(gilbert2d(width, height)):
-        array_2d[y][x] = array_1d[idx]
-
-    return array_2d
 
 def pad_sequence(sequence, max_length, pad_value, left_pad=False):
     actual_length = len(sequence)
@@ -196,6 +74,9 @@ def process_data(data_list):
             
             # Now process the actual test case
             test_input = np.array(test_example['input'])
+            dimensions = np.array(test_input.shape).reshape(1, 2)  
+            print(f"Test input shape: {test_input.shape}")  # Add this line to check the shape
+            
             test_input_flat = flatten_2d_to_1d(test_input)
             test_output_flat = flatten_2d_to_1d(np.array(test_example['output']))
 
@@ -203,10 +84,10 @@ def process_data(data_list):
             context = pad_sequence(context, MAX_CONTEXT_LENGTH, PAD_TOKEN, left_pad=True)
             target = pad_sequence([START_OUTPUT_MATRIX_TOKEN] + test_output_flat + [END_OUTPUT_MATRIX_TOKEN, END_SEQUENCE_TOKEN], MAX_PREDICTION_LENGTH, PAD_TOKEN)
             
-            dimensions = test_input.shape  # Store the dimensions of the input matrix
-            print(f"Context: {context}")
-            print(f"Target: {target}")
-            print(f"Dimensions: {dimensions}")  # This should now correctly print dimensions without error
+            dimensions = np.array(test_input.shape)
+            print(f"Processed data: Context: {context}")
+            print(f"Processed data: Target: {target}")
+            print(f"Processed data: Dimensions: {dimensions}")  # This should now correctly print dimensions without error
 
             # Append the tuple with dimensions
             processed_data.append((np.array(context), np.array(target), dimensions))
@@ -309,55 +190,6 @@ else:
 
 print("Data loading completed.")
 
-# Print a few lines to verify the data
-print("Training data examples:")
-for i in range(min(3, len(training_data))):
-    print(f"Input (length {len(training_data[i][0])}): {training_data[i][0]}")
-    print(f"Output (length {len(training_data[i][1])}): {training_data[i][1]}")
-    print("---")
-
-def test_mapping():
-    mapping_table = load_mapping_table()
-    for height in range(1, 31):
-        for width in range(1, 31):
-            mapping, encodings = mapping_table[(height, width)]
-            assert len(mapping) == height * width, f"Incorrect mapping length for {height}x{width}"
-            assert encodings.shape == (height, width, NUM_ENCODING_DIMENSIONS), f"Incorrect encoding shape for {height}x{width}: {encodings.shape}"
-            assert torch.all(encodings[:, :, 0] >= -1) and torch.all(encodings[:, :, 0] <= 1), "X values out of range"
-            assert torch.all(encodings[:, :, 1] >= -1) and torch.all(encodings[:, :, 1] <= 1), "Y values out of range"
-
-    print("All dimension and value tests passed.")
-
-
-    def test_case(matrix):
-        height, width = len(matrix), len(matrix[0]) if isinstance(matrix[0], list) else 1
-        flattened = flatten_2d_to_1d(matrix)
-
-        if height == 1 and width == 1:
-            reconstructed = [[flattened[0]]]
-        elif height == 1:
-            reconstructed = [flattened]
-        elif width == 1:
-            reconstructed = [[value] for value in flattened]
-        else:
-            reconstructed = [[0] * width for _ in range(height)]
-            for idx, value in enumerate(flattened):
-                x, y = mapping_table[(height, width)][0][idx]  # Get the x, y coordinates from the tuple
-                reconstructed[y][x] = value
-
-        assert matrix == reconstructed, f"Test failed for matrix: {matrix}"
-        print(f"Test passed for {height}x{width} matrix")
-
-
-
-    # Test cases
-    test_case([[1]])  # 1x1
-    test_case([[1, 2, 3]])  # 1x3
-    test_case([[1], [2], [3]])  # 3x1
-    test_case([[1, 2], [3, 4]])  # 2x2
-    test_case([[1, 2, 3], [4, 5, 6], [7, 8, 9]])  # 3x3
-    test_case([[i for i in range(j*5+1, (j+1)*5+1)] for j in range(5)])  # 5x5
-
 def analyze_dataset(data):
     """
     Analyze the dataset to find:
@@ -406,19 +238,6 @@ if __name__ == "__main__":
     ]
     training_data = load_and_process_training_data(training_file_paths)
     analyze_dataset(training_data)
-
-    mapping_table_file = 'mapping_table.pkl'
-    
-    if os.path.exists(mapping_table_file):
-        print("Loading existing mapping table...")
-        mapping_table = load_mapping_table(mapping_table_file)
-    else:
-        print("Creating new mapping table...")
-        mapping_table = create_mapping_table()
-        save_mapping_table(mapping_table, mapping_table_file)
-    
-    print("Running tests...")
-    test_mapping()
     
     test_data_1x1 = {
         "train": [{"input": [[1]], "output": [[0]]}],
