+++ hilbert_predictor/train.py
@@ -1,24 +1,25 @@
 import os
+import numpy
 import torch
 import torch.nn as nn
 import torch.optim as optim
 from torch.utils.data import DataLoader, TensorDataset
 from .model import (
     model,
-    checkpoint_path,
     d_model,
     nhead,
     num_layers,
     dim_feedforward,
-    MAX_CONTEXT_LENGTH,
-    MAX_PREDICTION_LENGTH,
-    dropout_rate,
     device,
     batch_size,
 )
+from .args import checkpoint_path, dropout_rate, batch_size
+
 from .data import (
     NUM_TOKENS,
     PAD_TOKEN,
+    MAX_CONTEXT_LENGTH,
+    MAX_PREDICTION_LENGTH,
     training_data,
 )
 from .args import args
@@ -27,6 +28,7 @@ from torch.nn import CrossEntropyLoss
 from collections import deque
 from typing import Dict, Optional, Literal
 from torch.nn.utils import clip_grad_norm
+
 def gradfilter_ma(
     m: nn.Module,
     grads: Optional[Dict[str, deque]] = None,
@@ -56,72 +58,79 @@ def gradfilter_ma(
     return grads
 
 def collate_fn(batch):
-    src_list, tgt_list = zip(*batch)
+    src_list, tgt_list, dimensions_list = zip(*batch)
+    # print(f"Batch dimensions: {dimensions_list}")
     src_padded = torch.nn.utils.rnn.pad_sequence(
-        [torch.tensor(src[:MAX_CONTEXT_LENGTH], dtype=torch.long) for src in src_list],
+        [src[:MAX_CONTEXT_LENGTH] for src in src_list],
         batch_first=True,
         padding_value=PAD_TOKEN,
     )
     tgt_padded = torch.nn.utils.rnn.pad_sequence(
         [
-            torch.tensor(tgt[:MAX_PREDICTION_LENGTH], dtype=torch.long)
+            tgt[:MAX_PREDICTION_LENGTH]
             for tgt in tgt_list
         ],
         batch_first=True,
         padding_value=PAD_TOKEN,
     )
-    src_lengths = torch.LongTensor(
+    src_lengths = torch.IntTensor(
         [min(len(src), MAX_CONTEXT_LENGTH) for src in src_list]
     )
-    tgt_lengths = torch.LongTensor(
+    tgt_lengths = torch.IntTensor(
         [min(len(tgt), MAX_PREDICTION_LENGTH) for tgt in tgt_list]
     )
-    return src_padded, tgt_padded, src_lengths, tgt_lengths
+    return src_padded, tgt_padded, src_lengths, tgt_lengths, dimensions_list
+
+def train_step(model, src, tgt, src_lengths, tgt_lengths, dimensions, criterion, teacher_forcing_ratio=1.0):
+    # print(f"train_step - src shape: {src.shape}")
+    # print(f"train_step - tgt shape: {tgt.shape}")
+    # print(f"train_step - dimensions: {dimensions}")
+
+    batch_size, seq_len = tgt.size()
+    outputs = torch.zeros(batch_size, seq_len, NUM_TOKENS + 1, device=device)
+
+    input_token = src[:, 0].unsqueeze(1)  # Ensure input_token is 2D
+    # print("Input token shape before model call:", input_token.shape)
+
+    for t in range(seq_len):
+        logits, confidences = model(input_token, dimensions)
+        logits, confidences = refine_predictions(model, input_token, logits, confidences, dimensions, threshold=0.5)
+        
+        outputs[:, t, :] = logits.squeeze(1)
+
+        if t < seq_len - 1:
+            use_teacher_forcing = torch.rand(1).item() < teacher_forcing_ratio
+            input_token = tgt[:, t+1].unsqueeze(1) if use_teacher_forcing else logits.argmax(-1).unsqueeze(1)
+
+    # Calculate loss excluding the last token of the target (which should be END_SEQUENCE_TOKEN)
+    total_loss = criterion(outputs[:, :-1, :].reshape(-1, NUM_TOKENS + 1), tgt[:, 1:].reshape(-1))
+    
+    # print(f"outputs shape: {outputs.shape}")
+    # print(f"tgt shape: {tgt.shape}")
+    # print(f"Loss calculation - outputs shape: {outputs[:, :-1, :].shape}, tgt shape: {tgt[:, 1:].shape}")
+
+    return outputs, total_loss
 
 
-def train_step(
-    model,
-    src,
-    tgt,
-    src_lengths,
-    tgt_lengths,
-    criterion,
-    train_loader,
-    teacher_forcing_ratio=1.0,
-):
-    with autocast(enabled=use_amp):
-        dimensions = src.size(0), src.size(1)  # Pass the dimensions of the input
-        logits = model(src, dimensions)  # Pass dimensions to the model
+def refine_predictions(model, src, initial_logits, confidences, dimensions, threshold=0.5):
+    _, initial_predictions = torch.max(initial_logits, dim=-1)
+    max_confidences, _ = torch.max(confidences, dim=-1)
+    low_confidence_mask = max_confidences < threshold
 
-        logits = logits[
-            :, : tgt.size(1), :
-        ]  # Ensure logits are the same length as targets
-
-        # Flatten for cross-entropy loss
-        logits = logits.reshape(
-            -1, NUM_TOKENS + 1
-        )  # Reshape to [batch_size * sequence_length, NUM_TOKENS + 1]
-        tgt = tgt.view(-1)  # Flatten target
-
-        # Calculate loss
-        loss = criterion(logits, tgt)
-
-    return logits, loss
+    refined_src = src.clone()
+    refined_src[low_confidence_mask] = initial_predictions[low_confidence_mask]
+    
+    # print(f"refined_src: {refined_src}")
+    
+    refined_logits, refined_confidences = model(refined_src, dimensions)
+    return refined_logits, refined_confidences
 
 
 if __name__ == "__main__":
     start_epoch = 0
     grads = None
-    accumulation_steps = 16  # Accumulate gradients over 16 batches
-    use_amp = True  # Use Automatic Mixed Precision
+    use_amp = False  # Use Automatic Mixed Precision
     print("Training the model.")
-    print("training_data", training_data[0])
-    # Prepare data loaders
-    train_dataset = TensorDataset(
-        torch.tensor([x[0] for x in training_data], dtype=torch.long),
-        torch.tensor([x[1] for x in training_data], dtype=torch.long),
-    )
-
     # Loss function and optimizer
     if os.path.exists(checkpoint_path):
         checkpoint = torch.load(checkpoint_path)
@@ -139,19 +148,22 @@ if __name__ == "__main__":
     start_epoch = 0
 
     criterion = CrossEntropyLoss(ignore_index=PAD_TOKEN)
+
     # Prepare data loaders
     train_dataset = TensorDataset(
-        torch.tensor([x[0] for x in training_data], dtype=torch.long),
-        torch.tensor([x[1] for x in training_data], dtype=torch.long),
+        torch.tensor(numpy.array([x[0] for x in training_data]), dtype=torch.long),
+        torch.tensor(numpy.array([x[1] for x in training_data]), dtype=torch.long),
+        torch.tensor(numpy.array([x[2] for x in training_data]), dtype=torch.long),  # Dimensions
     )
     train_loader = DataLoader(
         train_dataset, batch_size=batch_size, shuffle=True, collate_fn=collate_fn
     )
+
     # Training loop
     num_epochs = 10000
     model.train()
     
-    # needed for schedule-free optimizer
+    # Needed for schedule-free optimizer
     model.optimizer.train()
     
     if args.wandb:
@@ -161,7 +173,6 @@ if __name__ == "__main__":
             config={
                 "num_epochs": num_epochs,
                 "batch_size": batch_size,
-                "accumulation_steps": accumulation_steps,
                 "d_model": d_model,
                 "nhead": nhead,
                 "num_layers": num_layers,
@@ -175,13 +186,13 @@ if __name__ == "__main__":
         total_loss = 0
         model.optimizer.zero_grad()
 
-        for batch_idx, (src, tgt, src_lengths, tgt_lengths) in enumerate(train_loader):
+        for batch_idx, (src, tgt, src_lengths, tgt_lengths, dimensions) in enumerate(train_loader):
             src, tgt = src.to(device), tgt.to(device)
             src_lengths, tgt_lengths = src_lengths.to(device), tgt_lengths.to(device)
 
             with autocast(enabled=use_amp):
                 generated_ids, loss = train_step(
-                    model, src, tgt, src_lengths, tgt_lengths, criterion, train_loader
+                    model, src, tgt, src_lengths, tgt_lengths, dimensions, criterion
                 )
 
             loss.backward()
@@ -190,22 +201,16 @@ if __name__ == "__main__":
             # grads = gradfilter_ma(model, grads=grads)
             
             # Apply gradient clipping
-            # max_grad_norm = 1.0  # Adjust this value as needed
-            # clip_grad_norm(model.parameters(), max_grad_norm)
+            max_grad_norm = 1.0  # Adjust this value as needed
+            clip_grad_norm(model.parameters(), max_grad_norm)
 
-            if (batch_idx + 1) % accumulation_steps == 0:
-                model.optimizer.step()
-                model.optimizer.zero_grad()
+            model.optimizer.step()
+            model.optimizer.zero_grad()
             
-            # compute total batches
+            # Compute total batches
             print(f"Epoch {epoch+1}, Batch {batch_idx+1}, Loss: {loss.item():.4f}")
 
             total_loss += loss.item()
-
-        # If there are any accumulation steps left, do a final step
-        if (batch_idx + 1) % accumulation_steps != 0:
-            model.optimizer.step()
-            model.optimizer.zero_grad()
 
         avg_loss = total_loss / len(train_loader)
         print(f"Epoch {epoch+1} completed. Average Loss: {avg_loss:.4f}")
